+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to LazyProf will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Friendly error dialog when CraftLib dependency is missing
+- Graceful addon disable instead of silent failure
+
 ## [0.2.1] - 2026-01-23
 
 ### Added